+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - ZoomSettings now has options to control "RecordEnable", "DialOutEnable", and "MuteMyCameraOnStart"
  - ZoomSettings now has options to map Camera Devices to USB IDs
  - Added ActiveCamera property and event to IConnectProRoom
+ - Added event server messages for active camera changes
 
 ### Changed
  - Reworked camera presenters to allow selection and control from a single page